Shuffle around traits into more appropriate headers
