+++ include/albatross/src/cereal/variant.hpp
@@ -21,21 +21,15 @@ namespace cereal {
 
 namespace mapbox_variant_detail {
 
-template <typename Variant> struct variant_size {};
-
-template <typename... VariantTypes>
-struct variant_size<variant<VariantTypes...>>
-    : public std::tuple_size<std::tuple<VariantTypes...>> {};
-
 template <int N, class Variant, class... Args, class Archive,
-          typename std::enable_if_t<N == variant_size<Variant>::value, int> = 0>
+          typename std::enable_if_t<N == albatross::variant_size<Variant>::value, int> = 0>
 void load_variant(Archive &, int, Variant &) {
   assert(false); // load_variant received an out of bounds index.
 }
 
 template <
     int N, class Variant, class H, class... T, class Archive,
-    typename std::enable_if_t<N<variant_size<Variant>::value, int> = 0> void
+    typename std::enable_if_t<N<albatross::variant_size<Variant>::value, int> = 0> void
         load_variant(Archive &archive, int target, Variant &variant) {
   // It can get extremely confusing figuring out which type is causing a variant
   // to not be serializable, this static assert should help reveal the culprit.
