+++ tests/test_variant.cc
@@ -15,26 +15,5 @@
 
 namespace albatross {
 
-TEST(test_variant, test_details) {
-  const auto one =
-      cereal::mapbox_variant_detail::variant_size<variant<int>>::value;
-  EXPECT_EQ(one, 1);
-
-  const auto two =
-      cereal::mapbox_variant_detail::variant_size<variant<int, double>>::value;
-  EXPECT_EQ(two, 2);
-
-  struct X {};
-  const auto three = cereal::mapbox_variant_detail::variant_size<
-      variant<int, double, X>>::value;
-  EXPECT_EQ(three, 3);
-}
-
-TEST(test_variant, test_is_variant) {
-  EXPECT_TRUE(is_variant<variant<int>>::value);
-  EXPECT_TRUE(bool(is_variant<variant<int, double>>::value));
-  EXPECT_FALSE(is_variant<int>::value);
-  EXPECT_FALSE(is_variant<double>::value);
-}
 
 } // namespace albatross
